infoschema, executor: migrate dataForServersInfo from infoschema pkg to executor
